# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        example: student@example.com
        format: email
        type: string
    required:
    - email
    type: object
  dto.ForgotPasswordResponse:
    properties:
      message:
        example: если пользователь существует, письмо для сброса пароля отправлено
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: student@example.com
        format: email
        type: string
      password:
        example: StrongPass123
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      accessExpiresIn:
        example: 15m
        type: string
      message:
        example: успешный вход
        type: string
      refreshExpiresIn:
        example: 7d
        type: string
    type: object
  dto.LogoutAllResponse:
    properties:
      message:
        example: успешный выход из всех сессий
        type: string
    type: object
  dto.LogoutResponse:
    properties:
      message:
        example: успешный выход
        type: string
    type: object
  dto.OAuthCallbackResponse:
    properties:
      email:
        example: student@example.com
        type: string
      message:
        example: успешный вход через OAuth
        type: string
      userId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  dto.RefreshResponse:
    properties:
      accessExpiresIn:
        example: 15m
        type: string
      message:
        example: токены обновлены
        type: string
      refreshExpiresIn:
        example: 7d
        type: string
    type: object
...
  handler.AuthErrorResponse:
    properties:
      error:
        example: invalid_or_expired_token
        type: string
    type: object
...
host: localhost:4200
info:
  contact: {}
  description: 'REST API на Go (Gin + MongoDB + Redis + MinIO + RabbitMQ) с JWT/OAuth2-аутентификацией,
    CRUD-ресурсами, файловым хранилищем, асинхронной обработкой событий и Kubernetes-деплоем.
    ЛР9: health-зонды (/health/live, /health/ready), K8s-манифесты, горизонтальное
    масштабирование, Redis distributed lock.'
  title: Lab 2–9 REST API
  version: "1.4"
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email пользователя
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: сообщение о выполнении запроса
          schema:
            $ref: '#/definitions/dto.ForgotPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Запрос сброса пароля
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: успешный вход (JWT в HttpOnly cookies)
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Вход пользователя
      tags:
      - auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LogoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      security:
      - CookieAuth: []
      summary: Выход из текущей сессии
      tags:
      - auth
  /auth/logout-all:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LogoutAllResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      security:
      - CookieAuth: []
      summary: Выход из всех сессий
      tags:
      - auth
  /auth/oauth/{provider}:
    get:
      parameters:
      - description: Провайдер (yandex)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Редирект на провайдера
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Инициация OAuth авторизации
      tags:
      - auth
  /auth/oauth/{provider}/callback:
    get:
      parameters:
      - description: Провайдер (yandex)
        in: path
        name: provider
        required: true
        type: string
      - description: Код авторизации
        in: query
        name: code
        required: true
        type: string
      - description: State токен
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: успешный вход через OAuth (JWT в HttpOnly cookies)
          schema:
            $ref: '#/definitions/dto.OAuthCallbackResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Обработка OAuth callback
      tags:
      - auth
  /auth/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: успешное обновление (JWT в HttpOnly cookies)
          schema:
            $ref: '#/definitions/dto.RefreshResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Обновление токенов
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: успешная регистрация
          schema:
            $ref: '#/definitions/dto.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Токен и новый пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: пароль успешно изменён
          schema:
            $ref: '#/definitions/dto.ResetPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      summary: Установка нового пароля по токену
      tags:
      - auth
  /auth/whoami:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: данные текущего пользователя
          schema:
            $ref: '#/definitions/domain.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.AuthErrorResponse'
      security:
      - CookieAuth: []
      summary: Получение данных текущего пользователя
      tags:
      - auth
  /categories:
    get:
      parameters:
      - description: Номер страницы
        example: 1
        in: query
        name: page
        type: integer
      - description: Количество элементов на странице
        example: 10
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: список категорий
          schema:
            $ref: '#/definitions/dto.CategoryListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Список категорий
      tags:
      - categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: категория создана
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Создать категорию
      tags:
      - categories
  /categories/{id}:
    delete:
      parameters:
      - description: UUID категории
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: категория удалена
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Удалить категорию (soft delete)
      tags:
      - categories
    get:
      parameters:
      - description: UUID категории
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: категория найдена
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Получить категорию по UUID
      tags:
      - categories
    patch:
      consumes:
      - application/json
      parameters:
      - description: UUID категории
        in: path
        name: id
        required: true
        type: string
      - description: Тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PatchCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: категория обновлена
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Частичное обновление категории
      tags:
      - categories
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID категории
        in: path
        name: id
        required: true
        type: string
      - description: Тело запроса
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: категория обновлена
          schema:
            $ref: '#/definitions/dto.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apperror.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Полное обновление категории
      tags:
      - categories
  /files:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.FileResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Список файлов текущего пользователя
      tags:
      - files
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Файл
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UploadFileResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Загрузка файла
      tags:
      - files
  /files/{fileId}:
    delete:
      parameters:
      - description: ID файла
        in: path
        name: fileId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Удаление файла
      tags:
      - files
    get:
      parameters:
      - description: ID файла
        in: path
        name: fileId
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Скачивание файла по ID
      tags:
      - files
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"